pyproject.toml
[tool.poetry]
name = "btb-extraction"
version = "0.1.0"
description = ""
authors = ["sarrabenyahia <93934407+sarrabenyahia@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.1"
openpyxl = "^3.1.2"
pymupdf = "^1.23.25"
pyodbc = "^5.1.0"
chardet = ">=5.0"
psycopg2-binary = ">=2.9"
beautifulsoup4 = ">=4.12"
tqdm = ">=4.66"
python-dotenv = ">=1.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"
matplotlib = "^3.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
